+++ cmd/main.go
@@ -1,32 +1,32 @@
 package main
 
 import (
-	"context"
+	//"context"
 	"fmt"
-	"net/http"
+	//"net/http"
 	//"crypto/tls"
 	"github.com/amimof/huego/v2"
-	"encoding/json"
-	"io/ioutil"
+	//"encoding/json"
+	//"io/ioutil"
 )
 
-func errorHandler(res *http.Response) error {
-	var a *huego.APIResponse
-	if res.StatusCode >= 400 {
-		body, err := ioutil.ReadAll(res.Body)
-		if err != nil {
-			return err
-		}
-		err = json.Unmarshal(body, a)
-		if err != nil {
-			return err
-		}
-		for _, e := range a.Errors {
-			return e
-		}
-	}
-	return nil
-}
+// func errorHandler(res *http.Response) error {
+// 	var a *huego.APIResponse
+// 	if res.StatusCode >= 400 {
+// 		body, err := ioutil.ReadAll(res.Body)
+// 		if err != nil {
+// 			return err
+// 		}
+// 		err = json.Unmarshal(body, a)
+// 		if err != nil {
+// 			return err
+// 		}
+// 		for _, e := range a.Errors {
+// 			return e
+// 		}
+// 	}
+// 	return nil
+// }
 
 func main() {
 
@@ -78,29 +78,32 @@ func main() {
 
 
 	//Get one light
-	// light, err := c.GetLight("b1312d43-7aca-447b-92eb-95402aace153")
-	// if err != nil {
-	// 	panic(err)
-	// }
-	// fmt.Printf("ID: %s\nIDv1: %s\n", *light.ID, *light.IDv1)
-	// fmt.Printf("On: %t\n", *light.On.On)
-
-	// // Update the light
-	// ison := false
-	// light.On.On = &ison
-	// res, err := c.SetLightContext(context.Background(), *light.ID, light)
-	// if err != nil {
-	// 	panic(err)
-	// }
-	// fmt.Printf("Response: %s\n", string(res.BodyRaw))
-
-	//Get many lights
-	lights, err := c.GetLightsContext(context.Background())
+	light, err := c.GetLight("b1312d43-7aca-447b-92eb-95402aace153")
 	if err != nil {
 		panic(err)
 	}
-	fmt.Printf("Found %d lights\n", len(lights))
-	for i, light := range lights {
-		fmt.Printf("%d: %s\n", i, *light.ID)
+	fmt.Printf("ID: %s\nIDv1: %s\n", *light.ID, *light.IDv1)
+	fmt.Printf("On: %t\n", *light.On.On)
+
+	// Turning the light off
+	err = light.Off()
+	if err != nil {
+		panic(err)
 	}
+	// Turn the light on
+	err = light.TurnOn()
+	if err != nil {
+		panic(err)
+	}
+	
+
+	//Get many lights
+	// lights, err := c.GetLightsContext(context.Background())
+	// if err != nil {
+	// 	panic(err)
+	// }
+	// fmt.Printf("Found %d lights\n", len(lights))
+	// for i, light := range lights {
+	// 	fmt.Printf("%d: %s\n", i, *light.ID)
+	// }
 }
